@features/crud member

+++ src/components/participantContainer/ParticipantContainer.tsx
@@ -2,46 +2,80 @@ import React, {useEffect, useState} from 'react';
 import {View, Text, StyleSheet, Image, TouchableOpacity} from 'react-native';
 import firestore from '@react-native-firebase/firestore';
 import {useAppDispatch, useAppSelector} from '../../hooks/hook';
-import {addMember} from '../../features/memberSlice/member.slice';
+import {fetchMembers} from '../../features/allMembers/allMember.slice';
 import EditCreateParticipant from '../editCreateParticipant/EditCreateParticipant';
 import FontAwesome5 from 'react-native-vector-icons/FontAwesome5';
+import {deleteMember} from '../../features/deleteParticipantSlice/deleteParticipant.slice';
+import participantType from '../../types/participantType/participantType';
+import {
+  // BallIndicator,
+  // BarIndicator,
+  // DotIndicator,
+  // MaterialIndicator,
+  // PacmanIndicator,
+  // PulseIndicator,
+  // SkypeIndicator,
+  // UIActivityIndicator,
+  WaveIndicator,
+} from 'react-native-indicators';
 const ParticipantContainer = () => {
   const dispatch = useAppDispatch();
-  const memberData = useAppSelector(state => state.memberSlice);
-  const {data, loading, error} = memberData;
-  const [showOptions, setShowOptions] = useState<Boolean>(false);
+
+  //* fetch all member state
+  const allMember = useAppSelector(state => state.allMemberSlice);
+
+  //! create member state
+  const createParticipant = useAppSelector(state => state.createMemberSlice);
+
+  //? edit member state
+  const editMember = useAppSelector(state => state.editParticipantSlice.data);
+
+  //  delete member state
+  const deleteMembers = useAppSelector(
+    state => state.deleteParticipantSlice.deleteParticipant,
+  );
+
+  const {createdMember} = createParticipant;
+  const {allMembers} = allMember;
+  const [showOptions, setShowOptions] = useState<Boolean>(true);
   const [memberDelete, setMemberDelete] = useState<Boolean>(false);
   const [memberUpdate, setMemberUpdate] = useState<Boolean>(false);
   const [selectedMember, setSelectedMember] = useState<string | null>(null);
-  const [edit, setEdit] = useState(null);
+  const [edit, setEdit] = useState([]);
 
   //! Read operation from Member collection
   useEffect(() => {
     const allParticipants = async () => {
-      const member = await firestore().collection('Members').get();
-      const memberData = member.docs.map(items => {
-        console.warn(items.data());
+      const allMemberData = await firestore().collection('Members').get();
+      const memberData = allMemberData.docs.map(items => {
         return items.data();
       });
-      dispatch(addMember(memberData));
+      dispatch(fetchMembers(memberData));
     };
     allParticipants();
-  }, [addMember, dispatch]);
+  }, [dispatch, createdMember, editMember, deleteMembers]);
 
+  const toggleState = () => {
+    setShowOptions(false);
+  };
   //! Delete operation from Member collection
-  const handleDelete = async (documentId: string) => {
+  const handleDelete = async (documentId: string, data: participantType) => {
     try {
       setMemberDelete(true);
-      await firestore().collection('Members').doc(documentId).delete();
-      console.log('Document deleted successfully!', documentId);
-      setMemberDelete(false);
+      if (data.user_id === documentId) {
+        await firestore().collection('Members').doc(documentId).delete();
+        dispatch(deleteMember(data));
+        console.log('Document deleted successfully!', documentId);
+        setMemberDelete(false);
+      }
     } catch (error) {
       console.log('Error deleting document:', error);
     }
   };
 
   //! Update operation from Member collection
-  const handleUpdate = (documentId: string, data: any) => {
+
+  const handleUpdate = (documentId: string, data: participantType) => {
     if (data.user_id === documentId) {
       setSelectedMember(documentId);
       setEdit(data);
@@ -49,11 +83,13 @@ const ParticipantContainer = () => {
       console.log('Document clicked successfully!', documentId, data);
     } else console.log('not this document');
   };
-  const confirmUpdate = async (documentId, data) => {
+
+  //* confirm update function passed as props to edit component
+
+  const confirmUpdate = async (documentId: string, data: []) => {
     try {
       setMemberDelete(true);
       await firestore().collection('Members').doc(documentId).update(data);
-      // console.log('Document updated successfully!', documentId, data);
       console.log(data);
       setMemberDelete(false);
       setMemberUpdate(false);
@@ -62,7 +98,7 @@ const ParticipantContainer = () => {
     }
   };
 
-  const members = data?.map(member => {
+  const members = allMembers?.map(member => {
     return (
       <>
         <View
@@ -73,6 +109,8 @@ const ParticipantContainer = () => {
             justifyContent: 'space-between',
             marginVertical: 5,
             position: 'relative',
+            borderBottomWidth: 0.5,
+            paddingBottom: 12,
           }}>
           <View
             style={{
@@ -87,14 +125,20 @@ const ParticipantContainer = () => {
                 justifyContent: 'center',
                 alignItems: 'center',
               }}>
+              {showOptions && (
+                <View style={{width: '100%', height: '100%'}}>
+                  <WaveIndicator color="#8ac3ee" />
+                </View>
+              )}
               <Image
                 source={{uri: member.avarta}}
                 style={{width: '100%', height: '100%', borderRadius: 50}}
+                onLoad={toggleState}
               />
             </View>
             <View style={{marginLeft: 10}}>
-              <Text>{member.name}</Text>
-              <Text>{member.email}</Text>
+              <Text style={{fontSize: 18}}>{member.name}</Text>
+              <Text style={{fontSize: 16}}>{member.email}</Text>
             </View>
           </View>
           <View>
@@ -104,13 +148,13 @@ const ParticipantContainer = () => {
                   handleUpdate(member.user_id.toLocaleString(), member)
                 }>
                 <Text style={{marginRight: 8}}>
-                  <FontAwesome5 name="pen" color={'#fff'} size={25} />
+                  <FontAwesome5 name="pen" color={'#000'} size={25} />
                 </Text>
               </TouchableOpacity>
               <TouchableOpacity
-                onPress={() => handleDelete(member.user_id.toString())}>
+                onPress={() => handleDelete(member.user_id.toString(), member)}>
                 <Text style={{marginLeft: 8}}>
-                  <FontAwesome5 name="trash" color={'#fff'} size={25} />
+                  <FontAwesome5 name="trash" color={'#000'} size={25} />
                 </Text>
               </TouchableOpacity>
               {/* <View
@@ -137,21 +181,6 @@ const ParticipantContainer = () => {
                 }}></View> */}
             </TouchableOpacity>
           </View>
-          {showOptions && (
-            <View
-              style={{
-                // paddingVertical: 40,
-                paddingHorizontal: 15,
-                backgroundColor: '#fff',
-                width: '25%',
-                height: 100,
-                position: 'absolute',
-                right: -24,
-                top: -10,
-                justifyContent: 'space-evenly',
-                alignItems: 'center',
-              }}></View>
-          )}
         </View>
       </>
     );
